+++ game_client/Assets/Scripts/3_Main/SessionStatisticsPlain.cs
@@ -3,11 +3,11 @@ namespace _3_Main
     public class SessionStatisticsPlain
     {
         public int BulletsFired { get; set; }
-        public float DamageDone { get; set; }
+        public int DamageDone { get; set; }
         public int PelletsDestroyedSmall { get; set; }
         public int PelletsDestroyedMedium { get; set; }
         public int PelletsDestroyedLarge { get; set; }
-        public float Score { get; set; }
+        public int Score { get; set; }
         public int PowerUpBulletDamageCollected { get; set; }
         public int PowerUpBulletSpeedCollected { get; set; }
         public int PowerUpPlayerSpeedCollected { get; set; }
